rename interrupt_flags to io_registers
